Refactor import from folders
