+++ src/actions/stellar/index.ts
@@ -4,6 +4,7 @@ import sendTransaction from "./sendTransaction";
 import setKeypair from "./setKeypair";
 import paymentsHistory from "./paymentsHistory";
 import server from "./server";
+import resetKeys from "./resetKeys";
 
 export {
   minimumBalance,
@@ -12,4 +13,5 @@ export {
   setKeypair,
   paymentsHistory,
   server,
+  resetKeys,
 };
