+++ src/components/modals/home/StepTwoRegistration.tsx
@@ -4,10 +4,9 @@ import Input from "../../commons/Input";
 import { IRegistrationModalStepTwo } from "../../../types/types";
 import Link from "next/link";
 import { Checkbox } from "antd";
-import closeModalHandler from "@/actions/handlers/closeModalHandler";
 import { Modal } from "antd";
 import { handleCheck } from "../../../actions/utils";
-import loginHandler from "../../../actions/handlers/loginHandler";
+import { loginHandler, closeModalHandler } from "../../../actions/handlers";
 import { ILogin } from "../../../types/types";
 
 const StepTwoRegistration: React.FC<IRegistrationModalStepTwo> = ({
